fix: Do not depend on default vpc
For ease of use, this module will use the default vpc if no subnet
ids are passed as arguments but this make the module fail in case
there is no default vpc because it was deleted. This change
resolves this issue by adding counts on the data sources as well
and by making some black magic about tuples and lists.
There also some technical changes which include:
* deprectation notices
* pre-commit updates
* versions.tf file
* changelog
* author file